requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.16.0
django-cors-headers==4.4.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.2
idna==3.11
mysqlclient==2.2.8
oauthlib==3.3.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
xhtml2pdf==0.2.16
urllib3==2.6.3
